coder creation and destruction

+++ speex-safe/src/lib.rs
@@ -19,7 +19,7 @@ pub fn get_major_version() -> i32 {
     unsafe {
         let ptr = &mut major_version as *mut i32;
         let ptr = ptr as *mut c_void;
-        speex_lib_ctl(SPEEX_LIB_GET_MAJOR_VERSION as c_int, ptr);
+        speex_lib_ctl(SPEEX_LIB_GET_MAJOR_VERSION, ptr);
     }
     major_version
 }
@@ -29,7 +29,7 @@ pub fn get_minor_version() -> i32 {
     unsafe {
         let ptr = &mut minor_version as *mut i32;
         let ptr = ptr as *mut c_void;
-        speex_lib_ctl(SPEEX_LIB_GET_MINOR_VERSION as c_int, ptr);
+        speex_lib_ctl(SPEEX_LIB_GET_MINOR_VERSION, ptr);
     }
     minor_version
 }
@@ -39,7 +39,7 @@ pub fn get_micro_version() -> i32 {
     unsafe {
         let ptr = &mut micro_version as *mut i32;
         let ptr = ptr as *mut c_void;
-        speex_lib_ctl(SPEEX_LIB_GET_MICRO_VERSION as c_int, ptr);
+        speex_lib_ctl(SPEEX_LIB_GET_MICRO_VERSION, ptr);
     }
     micro_version
 }
@@ -48,7 +48,7 @@ pub fn get_extra_version() -> String {
     let cstr = unsafe {
         let mut str = null();
         let str_ptr = &mut str as *mut *const c_char;
-        speex_lib_ctl(SPEEX_LIB_GET_EXTRA_VERSION as c_int, str_ptr as *mut c_void);
+        speex_lib_ctl(SPEEX_LIB_GET_EXTRA_VERSION, str_ptr as *mut c_void);
         CStr::from_ptr(str)
     };
     cstr.to_string_lossy().into_owned()
@@ -58,10 +58,7 @@ pub fn get_version_string() -> String {
     let cstr = unsafe {
         let mut str = null();
         let str_ptr = &mut str as *mut *const c_char;
-        speex_lib_ctl(
-            SPEEX_LIB_GET_VERSION_STRING as c_int,
-            str_ptr as *mut c_void,
-        );
+        speex_lib_ctl(SPEEX_LIB_GET_VERSION_STRING, str_ptr as *mut c_void);
         CStr::from_ptr(str)
     };
     cstr.to_string_lossy().into_owned()
